Add VtsSecurityAvbTest
This commit adds VtsSecurityAvbTest that verifies system.
It loads contents and metadata of logical system partition, calculates
the hashtree from the contents, and compares with the metadata.

If the public key of GSI is given, the test case can be used to verify
that the system image is the GSI built by Google.

Bug: 65470881
Test: adb push gsi.avbpubkey /data/local/tmp
      adb shell /data/local/tmp/VtsSecurityAvbTest

Change-Id: Id98e67842e5c268f12e6132a2d2d86a3c9344d20