Check kernel version requirements for GRF devices.
... as required on the VSR documents. GRF devices must always
run a kernel that is higher than the one frozen for the
board API level.

The latest minlts frozen for a board API level can be found
under kernel/configs. For example, for Android 12, the latest one
in kernel/configs/s is 5.10; and
kernel/configs/s/android-5.10/android-base-conditional.xml specifies
also the minor version (5.10.43).

Test: TH
Bug: 255995840
Change-Id: I9e71186098cf7cca6f11ed75bc69f86710443b23