GKIComplianceTest: Add support for detecting if a device uses Android Go
Determining if a device uses Android Go will be required in the future
to ensure that Android Go devices only use GKI images released and
certified by Google. The approach that will be taken is to ensure
that Android Go devices only use GKI images that have been signed
by a specific set of keys.

Thus, move the logic that checks for different form factors into the
gsi_validation_utils.cpp file, and add support for detecting if a
device uses Android Go. The implementation of isGoDevice() is based
on the GMS Util library implementation.

Bug: 302020759
Change-Id: I0602377b3dc93c4f4fea727d17b46174441b46b9
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>