OGKI: Add functionality to get approved builds
Add a library method to parse the config for approved OGKI builds, and
return a mapping of build ids to build info for a given branch name.

BuildInfo struct is empty for now but it will be extended in the future
to represent information about approved OGKI builds.

The new library method will be used by the OGKI compliance VTS test.

Bug: 342094847
Test: adding new library method
Change-Id: I8580f20f6a2b2e9377487ef5bc6ce88c7b6ac49a